Update cache keys in `test-inputs.yml` (#79)
It's always caching. Or DNS. This time it was caching.

A while back, we changed the default cache key prefix from empty-string to `semver` so that our cache key names don't start with a dash, since users found that confusing.

We should have updated these tests accordingly at the time, but we did not. So how did the PR with the change pass?

It used the caches from older runs of the same job which used the old names. When those caches expired and were GC'd by GitHub, the jobs started failing.

This is what broke CI for #77. This PR will fix it.